Plugins: dependecy definition by plugin classes
The dependencies are now defined by the Plugin sub-class, opposed to a C
macro/generated function. This is possible because the dependency
information is only required when loading a plugin, and hence when the
Plugin is instantiated anyway. This has the beneficial property that
plugins can define their requirements based on the configuration at
run-time, and not only at compile-time. The PLUGIN_DEPENDS macro is no
more.