navgraph: provide NavGraphAspect, fix reloading bug
The plugin now registers as an aspect provider for the NavGraphAspect
allowing other plugins shared access to the TopologicalMapGraph
instance.

Along the way fix the long-standing issue that the graph was reloading
only once on the first change, but not anymore afterwards. Now properly
reloads the graph on each modification -- and only once per each
modification.