syncpoint: remove wait-for-one semantics of syncpoints
After discussing with Till we agree that we should make some adjustments
to the syncpoints model. As a first step, remove the capability to wait
for a single emitter on a syncpoint. Now, waiting on a syncpoint always
means to wait for all emitters to have fired.

We could only thing of one case where this may have provided some
benefit (think there are two producers of some data, and you want to
process it as soon as any new data is available, not only when all have
produced data). But that case seems a little far from actual use. And it
could be done on a plugin-level by having a set of threads waiting for a
number of inidivdual per-data syncpoints and synchronize this
internally.

The benefit of removing this capability is that it will make it much
easier to move to a modified, more declarative model of syncpoint and
thread interaction.